Adiconado site com teste.

+++ Running.Video/ConversionProgressEventArgs.cs
@@ -5,7 +5,7 @@ namespace Running.Video
   public class ConversionProgressEventArgs
   {
 
-    public ConversionProgressEventArgs(TimeSpan timeLeft, double progress, TimeSpan elapsedTime)
+    public ConversionProgressEventArgs(TimeSpan timeLeft, decimal progress, TimeSpan elapsedTime)
     {
       TimeLeft = timeLeft;
       Progress = progress;
@@ -13,7 +13,7 @@ namespace Running.Video
     }
     public TimeSpan ElapsedTime { get; private set; }
     public TimeSpan TimeLeft { get; private set; }
-    public double Progress { get; private set; }
+    public decimal Progress { get; private set; }
     //public string Description { get; set; }
   }
 }
